requirements.txt
aniso8601==8.1.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
Flask-RESTful==0.3.8
gensim==3.8.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.4
python-dotenv==0.15.0
pytz==2020.4
scipy==1.5.4
six==1.15.0
smart-open==4.0.1
Werkzeug==1.0.1
